pyproject.toml
[project]
name = "vasyan_core"
version = "0.1.0"
description = "TODO"
readme = "README.md"
authors = [
    { name = "Roman Karzhavin", email = "karzhavin.rd@ya.ru" }
]
requires-python = ">=3.13"
dependencies = [
    "fire>=0.7.1",
    "loguru>=0.7.3",
    "nats-py>=2.14.0",
    "pydantic>=2.12.5",
    "pyyaml>=6.0.3",
]

[build-system]
requires = ["uv_build>=0.8.22,<0.9.0"]
build-backend = "uv_build"
